% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_chrisg.R
\name{geom_chrisg}
\alias{geom_chrisg}
\title{geom_chrisg}
\usage{
geom_chrisg(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  inherit.aes = TRUE,
  na.rm = FALSE,
  by = "width",
  nudge_x = 0,
  ...
)
}
\arguments{
\item{mapping}{aes mapping}

\item{data}{data}

\item{stat}{stat}

\item{position}{position}

\item{inherit.aes}{logical, whether inherit aes from ggplot()}

\item{na.rm}{logical, whether remove NA values}

\item{by}{one of 'width' or 'height'}

\item{nudge_x}{horizontal adjustment to nudge Chris G}

\item{...}{additional parameters}
}
\value{
geom layer
}
\description{
geom layer adding chrisgs
}
\examples{
library("ggplot2")
ggplot(mtcars) +
ggchrisg::geom_chrisg(aes(mpg, wt), chrisg = "hagridhoodie", size = 5)

set.seed(1)
df <- data.frame(
	x = rnorm(10),
    y = rnorm(10),
    image = sample(c(
			"hagridhoodie",
        "fistbump",
        "madhatter",
        "poser",
        "conferencer"),
    size = 10,
    replace = TRUE))

ggplot(df) +
	geom_chrisg(aes(x, y, chrisg = image), size = 5)

}
